requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
requests
numpy
python-dotenv
google-generativeai
gunicorn
sentence-transformers
qdrant-client
transformers
PyPDF2
python-docx
huggingface-hub==0.14.1